new try

+++ hashmap.c
@@ -72,9 +72,9 @@ void enlarge(HashMap * map)
 {
   enlarge_called = 1; //no borrar (testing purposes)
   Pair** Auxiliar=map->buckets;
-  long CapacidadNueva = map->capacity*2;
+  int CapacidadNueva = map->capacity*2;
   map=createMap(CapacidadNueva);
-  for(long i=0;i<CapacidadNueva;i++)
+  for(int i=0;i<CapacidadNueva;i++)
   {
     if(Auxiliar[i] == NULL)
     {
